v24.9 - Batch/logging improvments. SSH retry logic (#58)
* Allways call return_result for exceptions thrown by batch task (#53)

* Move log level to 2nd column in logs (#55)

* Add retries to SSH connections (#56)

* Add missing changelog message

* Add tenacity for retries

* Add retry mechanism to SSH based transfers

* Update changelog

* Add a larger batch test (#57)

* bump version v24.8.0 -> v24.9.0

* Prevent local files being moved to staging directory (#59)

* Fix local transfer staging directory issue

* Don't delete the source file when copying locally

+++ src/opentaskpy/taskhandlers/transfer.py
@@ -71,7 +71,7 @@ class Transfer(TaskHandler):  # pylint: disable=too-many-instance-attributes
         if "OTF_STAGING_DIR" in environ:
             self.local_staging_dir = f"{environ['OTF_STAGING_DIR']}/{staging_dir_name}"
         else:
-            self.local_staging_dir = f"/{DEFAULT_STAGING_DIR_BASE}/{staging_dir_name}"
+            self.local_staging_dir = f"{DEFAULT_STAGING_DIR_BASE}/{staging_dir_name}"
 
         self.logger = opentaskpy.otflogging.init_logging(
             "opentaskpy.taskhandlers.transfer", self.task_id, TASK_TYPE
@@ -104,14 +104,25 @@ class Transfer(TaskHandler):  # pylint: disable=too-many-instance-attributes
                 self.logger.log(12, f"Closing dest connection for {remote_handler}")
                 remote_handler.tidy()
 
-        # Remove local staging directory if it exists
-        if path.exists(self.local_staging_dir):
+        # Remove local staging directory if it exists (and this isn't a local transfer)
+        if (
+            path.exists(self.local_staging_dir)
+            and self.local_staging_dir != self.source_file_spec["directory"]
+        ):
             self.logger.log(
                 12, f"Removing local staging directory {self.local_staging_dir}"
             )
             shutil.rmtree(self.local_staging_dir)
+        else:
+            self.logger.log(
+                12,
+                "Local staging directory is the same as source directory. Not removing",
+            )
 
         # Call super to do the rest
+        # Log the exception
+        if exception:
+            self.logger.exception(exception)
         return super().return_result(status, message, exception)  # type: ignore[no-any-return]
 
     def _get_default_class(self, protocol_name: str) -> type:
@@ -295,6 +306,9 @@ class Transfer(TaskHandler):  # pylint: disable=too-many-instance-attributes
                 else:
                     actual_sleep_seconds = sleep_seconds
 
+                # Prevent negative sleep times
+                actual_sleep_seconds = max(actual_sleep_seconds, 0)
+
                 self.logger.info(
                     f"No files found. Sleeping for {sleep_seconds} secs."
                     f" {remaining_seconds} seconds remain"
@@ -466,8 +480,11 @@ class Transfer(TaskHandler):  # pylint: disable=too-many-instance-attributes
                     )
                     or not self.source_remote_handler.supports_direct_transfer()
                 ):
-                    # Create local staging dir
-                    makedirs(self.local_staging_dir, exist_ok=True)
+                    # Create local staging dir (if this isn't using the local protocol)
+                    if self.source_file_spec["protocol"]["name"] != "local":
+                        makedirs(self.local_staging_dir, exist_ok=True)
+                    else:
+                        self.local_staging_dir = self.source_file_spec["directory"]
                     transfer_result = self.source_remote_handler.pull_files_to_worker(
                         remote_files, self.local_staging_dir
                     )
